# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
...
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "WISP Converter",
    "version" : "0.4.25"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/payments" : {
      "get" : {
        "operationId" : "redirect",
        "parameters" : [ {
          "example" : "identificativoIntermediarioPA_sessionId",
          "in" : "query",
          "name" : "idSession",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "302" : {
            "description" : "Redirect to Checkout service.",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "tags" : [ "Redirect" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/receipt" : {
      "get" : {
        "operationId" : "receiptRetrieve",
        "parameters" : [ {
          "in" : "query",
          "name" : "ci",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "ccp",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "iuv",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Receipt exists",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "tags" : [ "Receipt" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
    "/receipt/timer" : {
      "delete" : {
        "description" : "Delete a timer by paymentToken",
        "operationId" : "deleteTimer_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "paymentTokens",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully paymentToken expiration timer deleted",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "deleteTimer",
        "tags" : [ "ReceiptTimer" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Create a timer linked with paymentToken and receipt data",
        "operationId" : "createTimer_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReceiptTimerRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully paymentToken expiration timer created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "createTimer",
        "tags" : [ "ReceiptTimer" ]
      }
    },
...
    "/rpt/timer" : {
      "delete" : {
        "description" : "Delete a timer by sessionId",
        "operationId" : "deleteTimer",
        "parameters" : [ {
          "in" : "query",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully rpt timer deleted",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "deleteRPTTimer",
        "tags" : [ "RPTTimer" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Create a timer from sessionId data",
        "operationId" : "createTimer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RPTTimerRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully rpt timer created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "createRPTTimer",
        "tags" : [ "RPTTimer" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RPTTimerRequest" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string"
          }
        }
      },
...
      "ReceiptTimerRequest" : {
        "type" : "object",
        "properties" : {
          "expirationTime" : {
            "type" : "integer",
            "format" : "int64"
          },
          "fiscalCode" : {
            "pattern" : "\\w*",
            "type" : "string"
          },
          "noticeNumber" : {
            "pattern" : "\\d*",
            "type" : "string"
          },
          "paymentToken" : {
            "type" : "string"
          },
          "sessionId" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "ApiKey" : {
        "description" : "The API key to access this function app.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}